+++ src/domain/packages/validator/validations/all-mq-16/__tests__/data/ALL-MQ-16_success.json
