+++ src/domain/packages/validator/validations/all-mq-16/validate_ALL-MQ-16.ts
