+++ src/domain/packages/validator/validations/all-mq-19/__tests__/data/ALL-MQ-19_success.json
