+++ src/domain/packages/validator/validations/all-mq-19/validate_ALL-MQ-19.ts
