+++ src/domain/packages/validator/validations/all-mq-21/__tests__/data/ALL-MQ-21_fail_by_unexpected_symbol.json
