+++ src/domain/packages/validator/validations/all-mq-21/validate_ALL-MQ-21.ts
