+++ src/domain/packages/validator/validations/o-mq-2/__tests__/data/O-MQ-2_success.json
