+++ src/domain/packages/validator/validations/o-mq-2/validate_O-MQ-2.ts
