+++ src/domain/packages/validator/validations/og-mq-1/__tests__/data/OG-MQ-1_success.json
