+++ src/domain/packages/validator/validations/og-mq-1/__tests__/validate_OG-MQ-1.spec.ts
@@ -1,6 +1,6 @@
 import { validate_OG_MQ_1 } from "../validate-OG-MQ-1";
-import OG_MQ_1_success from "./data/OG_MQ_1_success.json";
-import OG_MQ_1_fail from "./data/OG_MQ_1_fail.json";
+import OG_MQ_1_success from "./data/OG-MQ-1_success.json";
+import OG_MQ_1_fail from "./data/OG-MQ-1_fail.json";
 
 describe("validate OG-MQ-1", () => {
     it("should not return errors for empty package", () => {
