+++ src/domain/packages/validator/validations/pr-st-3/__tests__/data/PR-ST-3_fail.json
