+++ src/domain/packages/validator/validations/pr-st-3/__tests__/validate_PR-ST-3.spec.ts
@@ -1,6 +1,6 @@
-import { validate_PR_ST_3 } from "../validate_PR_ST_3";
-import PR_ST_3_success from "./data/PR_ST_3_success.json";
-import PR_ST_3_fail from "./data/PR_ST_3_fail.json";
+import { validate_PR_ST_3 } from "../validate_PR-ST-3";
+import PR_ST_3_success from "./data/PR-ST-3_success.json";
+import PR_ST_3_fail from "./data/PR-ST-3_fail.json";
 
 describe("validate PR-ST-3", () => {
     it("should not return errors for empty package", () => {
