+++ src/domain/packages/validator/validations/pr-st-3/validate_PR-ST-3.ts
