+++ src/domain/packages/validator/validations/shst-mq-1/__tests__/data/SHST-MQ-1_fail.json
