+++ src/domain/packages/validator/validations/shst-mq-1/validate_SHST-MQ-1.ts
